adding a wrapper for tos
